+++ src/icons/ResumeIcon.svelte
@@ -1,0 +1,5 @@
+<svg viewBox="0 0 24 24" width="28" height="28" fill="currentColor">
+  <path
+    d="M14 2H6c-1.1 0-2 .9-2 2v16c0 1.1.9 2 2 2h12c1.1 0 2-.9 2-2V8l-6-6zm4 18H6V4h7v5h5v11zM8 15h8v2H8v-2zm0-3h8v2H8v-2z"
+  />
+</svg>
