+++ src/icons/index.js
@@ -4,3 +4,6 @@ export { default as ConciseIcon } from "./ConciseIcon.svelte";
 export { default as GridIcon } from "./GridIcon.svelte";
 export { default as MenuIcon } from "./MenuIcon.svelte";
 export { default as VerboseIcon } from "./VerboseIcon.svelte";
+export { default as GithubIcon } from "./GithubIcon.svelte";
+export { default as ResumeIcon } from "./ResumeIcon.svelte";
+export { default as LinkedInIcon } from "./LinkedInIcon.svelte";
