+++ styles.css
@@ -42,3 +42,97 @@ a {
   color: var(--link);
   text-decoration: underline;
 }
+
+/* Main layout */
+main {
+  position: relative;
+  z-index: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  max-height: 100vh;
+  padding: 0 20px;
+  box-sizing: border-box;
+}
+
+.card {
+  display: flex;
+  align-items: center;
+  gap: 1.5rem;
+}
+
+.photo {
+  width: 100px;
+  height: 100px;
+  object-fit: cover;
+  border-radius: 12px;
+}
+
+.info {
+  text-align: left;
+}
+
+.info h1 {
+  font-size: 1.8rem;
+  color: var(--text-secondary);
+  margin: 0 0 0.3rem;
+  font-weight: 500;
+}
+
+.tagline {
+  color: var(--text-faded);
+  font-size: 1.1rem;
+  margin: 0;
+}
+
+.links {
+  margin-top: 1.5rem;
+  display: flex;
+  gap: 1.5rem;
+}
+
+.links a {
+  color: var(--text-primary);
+  text-decoration: none;
+  transition: color 0.2s;
+}
+
+.links a:hover {
+  color: var(--text-secondary);
+}
+
+/* Mobile */
+@media (max-width: 600px) {
+  main {
+    justify-content: flex-start;
+    padding-top: 15vh;
+    gap: 1.2rem;
+  }
+
+  .card {
+    flex-direction: column;
+  }
+
+  .photo {
+    width: 80px;
+    height: 80px;
+  }
+
+  .info {
+    text-align: center;
+  }
+
+  .info h1 {
+    font-size: 1.4rem;
+  }
+
+  .tagline {
+    font-size: 0.95rem;
+  }
+
+  .links {
+    margin-top: 0.8rem;
+  }
+}
